+++ sporniket-javabeans-doclet/src/main/java/com/sporniket/libre/javabeans/doclet/expander/JavabeanGenerator.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package com.sporniket.libre.javabeans.doclet.encapsulator;
+package com.sporniket.libre.javabeans.doclet.expander;
 
 /**
  * @author dsporn
